server/requirements.txt
blinker==1.9.0
click==8.1.8
cloudpickle==3.1.1
dnspython==2.7.0
Flask==3.1.2
flask-cors==6.0.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
llvmlite==0.43.0
MarkupSafe==3.0.3
numba==0.60.0
numpy==2.0.2
packaging==25.0
pandas==2.3.3
pymongo==4.15.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
scikit-learn==1.6.1
scipy==1.13.1
shap==0.49.1
six==1.17.0
slicer==0.0.8
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
Werkzeug==3.1.3
zipp==3.23.0
